First attempt for improved attribute dumping

+++ src/main/java/org/stummi/jadis/element/ConstantPool.java
@@ -1,14 +1,14 @@
 package org.stummi.jadis.element;
 
-import org.stummi.jadis.element.constant.Constant;
-
 import lombok.Value;
+
+import org.stummi.jadis.element.constant.Constant;
 
 @Value
 public class ConstantPool implements Element {
 	private final Constant[] constants;
 
-	public Constant getConstant(short id) {
+	public Constant getConstant(int id) {
 		return constants[id];
 	}
 }
